third_party/blink/web_tests/http/tests/local/formdata/form-data-with-unknown-file-extension-expected.txt

PASS xhr.responseText is 'file1=file.invalid:application/octet-stream:1234567890'
PASS successfullyParsed is true

TEST COMPLETE
